+++ tests/text/test_wer.py
@@ -1,83 +1,75 @@
+from typing import Callable, List, Union
+
 import pytest
 
+from tests.text.helpers import INPUT_ORDER, TextTester
 from torchmetrics.utilities.imports import _JIWER_AVAILABLE
 
 if _JIWER_AVAILABLE:
     from jiwer import compute_measures
+else:
+    compute_measures = Callable
 
 from torchmetrics.functional.text.wer import wer
 from torchmetrics.text.wer import WER
 
+BATCHES_1 = {"preds": [["hello world"], ["what a day"]], "targets": [["hello world"], ["what a wonderful day"]]}
 
-@pytest.mark.parametrize(
-    "hyp,ref,score",
-    [(["hello world"], ["hello world"], 0.0), (["Firwww"], ["hello world"], 1.0)],
-)
+BATCHES_2 = {
+    "preds": [
+        ["i like python", "what you mean or swallow"],
+        ["hello duck", "i like python"],
+    ],
+    "targets": [
+        ["i like monthy python", "what do you mean, african or european swallow"],
+        ["hello world", "i like monthy python"],
+    ],
+}
+
+
+def _compute_wer_metric_jiwer(prediction: Union[str, List[str]], reference: Union[str, List[str]]):
+    return compute_measures(reference, prediction)["wer"]
+
+
 @pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_same(hyp, ref, score):
-    """Test to ensure that the torchmetric WER matches reference scores."""
-    metric = WER()
-    metric.update(hyp, ref)
-    assert metric.compute() == score
-
-
 @pytest.mark.parametrize(
-    "hyp,ref,expected_score,expected_incorrect,expected_total",
+    ["preds", "targets"],
     [
-        (["hello world"], ["hello world"], 0.0, 0, 2),
-        (["Firwww"], ["hello world"], 1.0, 2, 2),
+        pytest.param(BATCHES_1["preds"], BATCHES_1["targets"]),
+        pytest.param(BATCHES_2["preds"], BATCHES_2["targets"]),
     ],
 )
-@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_functional(ref, hyp, expected_score, expected_incorrect, expected_total):
-    """Test to ensure that the torchmetric functional WER matches the jiwer reference."""
-    assert wer(ref, hyp) == expected_score
+class TestWER(TextTester):
+    @pytest.mark.parametrize("ddp", [False, True])
+    @pytest.mark.parametrize("dist_sync_on_step", [False, True])
+    def test_wer_class(self, ddp, dist_sync_on_step, preds, targets):
 
+        self.run_class_metric_test(
+            ddp=ddp,
+            preds=preds,
+            targets=targets,
+            metric_class=WER,
+            sk_metric=_compute_wer_metric_jiwer,
+            dist_sync_on_step=dist_sync_on_step,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )
 
-@pytest.mark.parametrize(
-    "hyp,ref",
-    [
-        (["hello world"], ["hello world"]),
-        (["Firwww"], ["hello world"]),
-    ],
-)
-@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_reference_functional(hyp, ref):
-    """Test to ensure that the torchmetric functional WER matches the jiwer reference."""
-    assert wer(ref, hyp) == compute_measures(ref, hyp)["wer"]
+    def test_wer_functional(self, preds, targets):
 
+        self.run_functional_metric_test(
+            preds,
+            targets,
+            metric_functional=wer,
+            sk_metric=_compute_wer_metric_jiwer,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )
 
-@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_reference_functional_concatenate():
-    """Test to ensure that the torchmetric functional WER matches the jiwer reference when concatenating."""
-    ref = ["hello world", "hello world"]
-    hyp = ["hello world", "Firwww"]
-    assert wer(ref, hyp) == compute_measures(ref, hyp)["wer"]
-    assert wer(hyp, ref, concatenate_texts=True) == compute_measures("".join(ref), "".join(hyp))["wer"]
+    def test_wer_differentiability(self, preds, targets):
 
-
-@pytest.mark.parametrize(
-    "hyp,ref",
-    [
-        (["hello world"], ["hello world"]),
-        (["Firwww"], ["hello world"]),
-    ],
-)
-@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_reference(hyp, ref):
-    """Test to ensure that the torchmetric WER matches the jiwer reference."""
-    metric = WER()
-    metric.update(hyp, ref)
-    assert metric.compute() == compute_measures(ref, hyp)["wer"]
-
-
-@pytest.mark.skipif(not _JIWER_AVAILABLE, reason="test requires jiwer")
-def test_wer_reference_batch():
-    """Test to ensure that the torchmetric WER matches the jiwer reference with accumulation."""
-    batches = [("hello world", "Firwww"), ("hello world", "hello world")]
-    metric = WER()
-
-    for hyp, ref in batches:
-        metric.update(ref, hyp)
-    reference_score = compute_measures(truth=[x[0] for x in batches], hypothesis=[x[1] for x in batches])["wer"]
-    assert metric.compute() == reference_score
+        self.run_differentiability_test(
+            preds=preds,
+            targets=targets,
+            metric_module=WER,
+            metric_functional=wer,
+            input_order=INPUT_ORDER.PREDS_FIRST,
+        )